Armor equipment system

+++ Assets/Scripts/Components/HeadArmor.cs
@@ -2,6 +2,6 @@
 {
     public struct HeadArmor
     {
-        public float Armor;
+        public int Armor;
     }
 }
